% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_count.R
\name{as_count}
\alias{as_count}
\alias{as_count<-}
\title{Set Pretty/Count Printing of a term_count Object}
\usage{
as_count(x, value = FALSE)

as_count(x) <- value
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{value}{logical.  If \code{TRUE} the object will attempt to be printed
pretty.}
}
\value{
Returns a \code{term_count} with the \code{pretty}
\code{\link[base]{attributes}} set.
}
\description{
\code{as_count} - Set the pretty printing of a \code{term_count} object.
Either print pretty as a combination of count and percent/proportion or as
just counts.

\code{pretty<-} - Set the pretty/count printing of a \code{term_count} object.
}
\details{
Note that pretty printing can be turned off globally by setting
\code{options(termco_pretty = FALSE)}.
}
\examples{
out <- as_count(markers, FALSE)
out
as_count(out) <- TRUE
out
}
\seealso{
\code{\link[termco]{as_count}}
}
\keyword{count}
\keyword{pretty}
\keyword{printing}
