% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dtm.R
\name{as_dtm}
\alias{as_dtm}
\alias{as_tdm}
\title{Convert a \code{term_count} Object Into a \code{DocumentTermMatrix}/\code{TermDocumentMatrix}}
\usage{
as_dtm(x, weighting = tm::weightTf, ...)

as_tdm(x, weighting = tm::weightTf, ...)
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{weighting}{A weighting function capable of handling a \code{tm::DocumentTermMatrix}. It defaults to \code{weightTf} for term frequency weighting. Available weighting functions shipped with the \pkg{tm} package are \code{weightTf}, \code{weightTfIdf}, \code{weightBin}, and \code{weightSMART}.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{tm::DocumentTermMatrix} or \code{tm::TermDocumentMatrix} object.
}
\description{
Convenience functions to convert a \code{term_count} object into either a \code{tm::DocumentTermMatrix} or \code{tm::TermDocumentMatrix} object.  Grouping variables are used as the row/column names for the \code{DocumentTermMatrix}/\code{TermDocumentMatrix}.
}
\examples{
as_dtm(markers)
as_dtm(markers,weighting = tm::weightTfIdf)
as_tdm(markers)

cosine_distance <- function (x, ...) {
    x <- t(slam::as.simple_triplet_matrix(x))
    stats::as.dist(1 - slam::crossprod_simple_triplet_matrix(x)/(sqrt(slam::col_sums(x^2) \%*\%
        t(slam::col_sums(x^2)))))
}


mod <- hclust(cosine_distance(as_dtm(markers)))
plot(mod)
rect.hclust(mod, k = 5, border = "red")

(clusters <- cutree(mod, 5))
}
\keyword{documenttermmatrix,}
\keyword{termdocumentmatrix}

