% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_terms.R
\name{as_terms}
\alias{as_terms}
\title{Convert a Count Matrix to List of Term Vectors}
\usage{
as_terms(x, names = NULL, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}}/\code{\link[base]{matrix}} of counts.}

\item{names}{A character vector of names to assign to the list.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of term vectors.
}
\description{
Convert a count matrix to a named list of term vectors.
}
\examples{
data(markers)
as_terms(markers)
}
