% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_project.R
\name{classification_project}
\alias{classification_project}
\alias{is.global}
\title{Classification Project Template}
\usage{
classification_project(path = "new", open = is.global(2))

is.global(n = 1)
}
\arguments{
\item{path}{The path (including project name) for the project.}

\item{open}{logical. If \code{TRUE} the project will be opened in RStudio.
The default is to test if new_project is being used in the global environment,
if it is then the project directory will be opened.}

\item{n}{The number of generations to go back. If used as a function argument
\code{n} should be set to 2.}
}
\description{
Create a template directory of subdirectories and files for a classification
project.
}
\examples{
\dontrun{
classification_project()
}
}
\keyword{project}

