% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_template.R
\name{classification_template}
\alias{classification_template}
\title{Generate a Basic Classification Analysis Template}
\usage{
classification_template(path = "scripts/02_classification.R",
  file.ext = "fst", categories.file = "models/categories.R",
  verbose = TRUE, ...)
}
\arguments{
\item{path}{Path to classification template script.}

\item{file.ext}{A supported file extension for the text data.  This dictates
the packages used in the template for reading in data.  Currently,
'.csv' (\pkg{readr})and '.fst' (\pkg{fst}) are supported.  This can be
manually changed after the template is created.}

\item{categories.file}{A path to the categories model (term list) file.  Defaults
to \file{'models/categories.R'}.}

\item{verbose}{logical.  If \code{TRUE} a message is printed if the template
inclusion was successful.}

\item{\ldots}{ignored.}
}
\description{
Generates a basic template for a classification using \pkg{termco} tools.
}
\examples{
\dontrun{
classification_template()
}
}
