% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify Rows of a Count Matrix}
\usage{
classify(x, n = 1, ties.method = "random", ...)
}
\arguments{
\item{x}{A \code{term_count} or count \code{\link[base]{matrix}}/\code{\link[base]{data.frame}}.}

\item{n}{The number of classifications per row to return.}

\item{ties.method}{Either \code{c("random", "first", "last")} for specifying
how ties are handled; "random" by default.  See \code{\link[base]{max.col}}.}

\item{\ldots}{ignored.}
}
\value{
Returns a single \code{\link[base]{vector}} or \code{\link[base]{list}}
of ordered vectors of predicted classifications; order by term frequency.
Ties default to random order.
}
\description{
Use a count matrix to classify the rows, based on the frequencies in the cells.
}
\examples{
data(markers)
classify(markers)
classify(markers, n = 2)
}
\seealso{
\code{\link[base]{max.col}}
}
\keyword{classify}
\keyword{predict}

