% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colo.R
\name{colo}
\alias{colo}
\title{Make Regex to Locate Strings Containing Coocuring Substrings}
\usage{
colo(..., not = NULL, copy2clip = getOption("termco.copy2clip"))
}
\arguments{
\item{not}{A substring to exclude from consideration.}

\item{copy2clip}{logical.  If code{TRUE} uses \code{\link[clipr]{write_clip}}
to copy the output to the clipboard.}

\item{\ldots}{Terms that cooccur/colocate}
}
\value{
Returns a regular expression.  If Windows attempts to copy to
clipboard as well.
}
\description{
Make a regex to locate strings that contain >= 2 substrings with optional
negation.
}
\examples{
colo('overall', 'course')
colo('overall', 'course', "eval")
colo('overall', 'course', not="instructor")

\dontrun{
search_term(sam_i_am, colo("^i\\\\b", "like"))
search_term(sam_i_am, colo("^i\\\\b", "like", "not"))
search_term(sam_i_am, colo("^i\\\\b", "like", not="not")
}
}
\keyword{colocate}
\keyword{cooccur}

