% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_counts.R
\name{combine_counts}
\alias{combine_counts}
\title{Combine \code{term_count} and \code{token_count} Objects}
\usage{
combine_counts(x, y, mapping = NULL, ...)
}
\arguments{
\item{x}{A \code{term_count} or \code{token_count} object.}

\item{y}{A \code{term_count} or \code{token_count} object.}

\item{mapping}{A list of named vectors where the vector names are the collapsed
column names and the vectors are the names of the columns to collapse.  The
default, \code{NULL}, combines columns with the same name in both \code{x} and
\code{y}.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{combine_counts} object.
}
\description{
Combine \code{term_count} and \code{token_count} objects.
}
\examples{
token_list <- list(
    list(
        person = c('sam', 'i')
    ),
    list(
        place = c('here', 'house'),
        thing = c('boat', 'fox', 'rain', 'mouse', 'box', 'eggs', 'ham')
    ),
    list(
        no_like = c('not like'),
        thing = c('train', 'goat')
    )
)

(y <- token_count(sam_i_am, grouping.var = TRUE, token.list = token_list))


term_list <- list(
    list(Is = c("I")),
    list(
        oulds = c("ould"),
        thing = c('egg', 'ham')
    )
)


(x <-term_count(sam_i_am, grouping.var = TRUE, term_list, ignore.case = FALSE))

combine_counts(x, y)
combine_counts(y, x)

library(lexicon)
library(textshape)
library(dplyr)

token_list <- lexicon::nrc_emotions \%>\%
    textshape::column_to_rownames() \%>\%
    t() \%>\%
    textshape::as_list()


a <- presidential_debates_2012 \%>\%
     with(token_count(dialogue, list(person, time), token_list))

term_list <- list(
    response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
)

b <- presidential_debates_2012 \%>\%
     with(term_count(dialogue, list(person, time), term_list))

combine_counts(a, b)
combine_counts(b, a)


d <- sam_i_am \%>\%
     term_count(TRUE, token_list[1:2])

e <- sam_i_am \%>\%
     term_count(TRUE, token_list[[3]])


combine_counts(e, d)
}
