% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Extract Coverage Information}
\usage{
coverage(x, ...)
}
\arguments{
\item{x}{A \code{\link[termco]{term_count}} object.}

\item{\ldots}{ignored}
}
\value{
Returns a list:
\item{not}{A logical vector of all rows not covered (row sums equal zero)}
\item{covered}{A logical vector of all rows covered (row sums greater than zero)}
\item{coverage}{The percentage rate of \deqn{\frac{covered}{not + covered}}{covered/(not + covered)}}
\item{n_covered}{The row sums of the unique terms}
\item{total_terms}{The row sums of the terms}
}
\description{
Extract coverage information from a termco object including the percentage of
rows that sum to zero as well as the location of non-covered rows for easy
extraction.
}
\examples{
data(pres_debates2012)

discoure_markers <- list(
    like = c("love", "like"),
    water = c("lake", "ocean", "water"),
    justify = c("because"),
    he = c("\\\\bhe", "him"),
    we = c("\\\\bwe", "\\\\bus", "\\\\bour")
)

library(dplyr)
(markers2 <- with(dplyr::mutate(pres_debates2012, turn = dplyr::id(dialogue)),
    term_count(dialogue, turn, discoure_markers)
))

coverage(markers2)

pres_debates2012[coverage(markers2)$not, "dialogue"] \%>\%
   c()
}
\keyword{coverage}

