% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\alias{get_text.term_count}
\title{Get a Text Stored in Various Objects}
\usage{
get_text(x, ...)

\method{get_text}{term_count}(x, ...)
}
\arguments{
\item{x}{A \code{\link[termco]{term_count}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector or list of text strings.
}
\description{
Extract the text supplied to the
\code{\link[termco]{term_count}} object.
}
\examples{
library(dplyr)

discoure_markers <- list(
    response_cries = c("\\\\boh\\\\b", "\\\\bah\\\\b", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "\\\\bhey",
    justification = "because"
)

model <- presidential_debates_2012 \%>\%
    with(term_count(dialogue, grouping.var = TRUE, discoure_markers))

get_text(model, 'summons')
get_text(model, 'response_cries')
get_text(model, c('summons', 'response_cries'))
}

