% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage.R
\name{hierarchical_coverage}
\alias{hierarchical_coverage}
\title{Hierarchical Coverage}
\usage{
hierarchical_coverage(x, terms, bind = TRUE, ignore.case = TRUE, ...)
}
\arguments{
\item{x}{A text vector (vector of strings).}

\item{terms}{A vector of regular expressions to match against \code{x}.}

\item{bind}{logical.  If \code{TRUE} The terms are bound with boundary
markers to ensure \code{"read"} matches \code{"read"} but not \code{"ready"}).}

\item{ignore.case}{logical.  Should case be ignored in matching the
\code{terms} against \code{x}?}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{\link[base]{data.frame}} with 3 columns:
\describe{
  \item{terms}{the search term}
  \item{unique}{the unique coverage of the term}
  \item{cumulative}{the cumulative coverage of the term}
}
}
\description{
The unique coverage of a text vector by a term after partitioning out the
elements matched by previous terms.
}
\examples{
x <- presidential_debates_2012[["dialogue"]]
terms <- frequent_terms(x)[[1]]
(out <- hierarchical_coverage(x, terms))
plot(out)

(out2 <- hierarchical_coverage(x, frequent_terms(x, 30)[[1]]))
plot(out2, use.terms = TRUE)
plot(out2, use.terms = TRUE, mark.one = FALSE)
}
\author{
Steve T. Simpson and Tyler Rinker <tyler.rinker@gmail.com>.
}
\keyword{coverage}

