% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage_term.R
\name{hierarchical_coverage_term}
\alias{hierarchical_coverage_term}
\title{Hierarchical Coverage of Terms}
\usage{
hierarchical_coverage_term(text.var, terms, bound = TRUE,
  ignore.case = TRUE, sort = FALSE, ...)
}
\arguments{
\item{text.var}{A text vector (vector of strings).}

\item{terms}{A vector of regular expressions to match against \code{x}.}

\item{bound}{logical.  If \code{TRUE} the terms are bound with boundary
markers to ensure \code{"read"} matches \code{"read"} but not \code{"ready"}).}

\item{ignore.case}{logical.  Should case be ignored in matching the
\code{terms} against \code{x}?}

\item{sort}{logical.  If \code{TRUE} the output is sorted by highest unique
gain.  If \code{FALSE} order of term input is retained.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{\link[base]{data.frame}} with 3 columns:
\describe{
  \item{terms}{the search term}
  \item{unique}{the unique coverage of the term}
  \item{cumulative}{the cumulative coverage of the term}
}
}
\description{
The unique coverage of a text vector by a term after partitioning out the
elements matched by previous terms.
}
\examples{
x <- presidential_debates_2012[["dialogue"]]
terms <- frequent_terms(x)[[1]]
(out <- hierarchical_coverage_term(x, terms))
plot(out)

(out2 <- hierarchical_coverage_term(x, frequent_terms(x, 30)[[1]]))
plot(out2, use.terms = TRUE)
plot(out2, use.terms = TRUE, mark.one = TRUE)
}
\author{
Steve T. Simpson and Tyler Rinker <tyler.rinker@gmail.com>.
}
\seealso{
Other hierarchical_coverage functions: \code{\link{hierarchical_coverage_regex}}
}
\keyword{coverage}

