% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_terms.R
\name{important_terms}
\alias{important_terms}
\title{Top Min-Max Scaled TF-IDF terms}
\usage{
important_terms(text.var, n = 20, stopwords = tm::stopwords("en"),
  stem = FALSE, language = "porter", strip = TRUE,
  strip.regex = "[^A-Za-z' ]", ...)
}
\arguments{
\item{text.var}{A vector of character strings.}

\item{n}{The number of rows to print.  If integer selects the frequency at
the nth row and prints all rows >= that value.  If proportional (less than 0)
the frequency value for the nth\% row is selected and prints all rows >= that
value.}

\item{stopwords}{A vector of stopwords to exclude.}

\item{stem}{logical.  If \code{TRUE} the \code{\link[SnowballC]{wordStem}}
is used with \code{language = "porter"} as the default.  Note that stopwords
will be stemmed as well.}

\item{language}{The stem language to use (see  \code{\link[SnowballC]{wordStem}}).}

\item{strip}{logical.  If \code{TRUE} all values that are not alpha, apostrophe,
or spaces are stripped.  This regex can be changed via the \code{strip.regex}
argument.}

\item{strip.regex}{A regular expression used for stripping undesired characters.}

\item{\ldots}{\code{\link[gofastr]{remove_stopwords}}}
}
\value{
Returns a \code{\link[base]{data.frame}} of terms and min-max scaled tf-idf weights.
}
\description{
View the top n min-max scaled tf-idf weighted terms in a text.
}
\examples{
\dontrun{
x <- presidential_debates_2012[["dialogue"]]

frequent_terms(x)
important_terms(x)
important_terms(x, n=899)
important_terms(x, n=.1)
important_terms(x, min.char = 7)
important_terms(x, min.char = 6, stem=TRUE)

plot(important_terms(x))
plot(important_terms(x, n = .02))
plot(important_terms(x, n = 40))
plot(important_terms(x, n = 100), as.cloud = TRUE)
}
}
\keyword{important}
