% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termco-package.R
\docType{data}
\name{markers}
\alias{markers}
\title{Discourse Marker Search of Presidential Debates}
\format{A data frame with 10 rows and 7 variables}
\usage{
data(markers)
}
\description{
A \code{\link[termco]{term_count}} dataset containing discourse markers from
the 2012 presidential debates.
}
\details{
\itemize{
  \item person. The speaker
  \item time. Variable indicating which of the three debates the dialogue is from
  \item n.words. The number of words
  \item response_cries. The number of response cries: \code{c("oh", "ah", "aha", "ouch", "yuk")}
  \item back_channels. The number of back channels: \code{c("uh[- ]huh", "uhuh", "yeah")}
  \item summons. The number of summons: \code{"hey"}
  \item justification. The number of justification: \code{"because"}
}
}
\keyword{datasets}
