% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_word.R
\name{match_word}
\alias{match_word}
\title{Regular Expression Matching of Words in a Text}
\usage{
match_word(text.var, term.list, ...)
}
\arguments{
\item{text.var}{The text string variable.}

\item{term.list}{A list of named character vectors.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of matched word vectors.
}
\description{
Extract the words from a text that match a regular expression.
}
\examples{
match_word(
    text.var = c(
        'the dog is todo them funny',
        'I\\'d wait to eat the sandwich.',
        'the dog ate the sandwiches'
    ),
    term.list = c('the', 'sandwich', 'do', '^do', '\\\\bdo', 'do$')
)

match_word(
    text.var = c(
        'the dog is todo them funny',
        'I\\'d wait to eat the sandwich.',
        'the dog ate the sandwiches'
    ),
    term.list = list(
        'the',
        'sandwich',
        'do',
        '^do',
        '\\\\bdo',
        'do$',
        '\\\\b(eat|ate)\\\\b'
    )
)
}
