% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{plot.classify}
\alias{plot.classify}
\title{Plots a plot.classify Object}
\usage{
\method{plot}{classify}(x, rm.na = TRUE, ...)
}
\arguments{
\item{x}{A classify object.}

\item{rm.na}{logical.  If \code{TRUE} unclassified \code{NA} values are not displayed.}

\item{\ldots}{Other arguments passed to \code{\link[termco]{plot_counts}}.}
}
\description{
Plots a plot.classify object
}

