% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_ngrams.R
\name{plot.frequent_ngrams}
\alias{plot.frequent_ngrams}
\title{Plots a frequent_ngrams Object}
\usage{
\method{plot}{frequent_ngrams}(x, drop.redundant.yaxis.text = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{The \code{frequent_ngrams} object.}

\item{drop.redundant.yaxis.text}{logical.  If \code{TRUE} the second y axis text/ticks,
in the heat plot are dropped.}

\item{plot}{logical.  If \code{TRUE} the output is plotted.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of the three \pkg{ggplot2} objects that make the
combined plot.
}
\description{
Plots a frequent_ngrams object.
}
