% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage.R
\name{plot.hierarchical_coverage}
\alias{plot.hierarchical_coverage}
\title{Plots a hierarchical_coverage Object}
\usage{
\method{plot}{hierarchical_coverage}(x, use.terms = nrow(x) <= 30,
  mark.one = TRUE, ...)
}
\arguments{
\item{x}{A hierarchical_coverage object.}

\item{use.terms}{logical.  If \code{TRUE} terms are plotted on the x axis.
If \code{FALSE} word numbers are.  Te default is to plot terms if they are
equal to or less than 30 in length.}

\item{mark.one}{logical.  If \code{TRUE} a purple horizontal line is added at
100\% and the y axis is extended as well.}

\item{\ldots}{ignored.}
}
\description{
Plots a hierarchical_coverage object
}

