% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage_regex.R
\name{plot.hierarchical_coverage_regex}
\alias{plot.hierarchical_coverage_regex}
\title{Plots a hierarchical_coverage_regex Object}
\usage{
\method{plot}{hierarchical_coverage_regex}(x, use.names = nrow(x) <= 30,
  mark.one = FALSE, ...)
}
\arguments{
\item{x}{A hierarchical_coverage_regex object.}

\item{use.names}{logical.  If \code{TRUE} terms are plotted on the x axis.
If \code{FALSE} word numbers are.  Te default is to plot terms if they are
equal to or less than 30 in length.}

\item{mark.one}{logical.  If \code{TRUE} a purple horizontal line is added at
100\% and the y axis is extended as well.}

\item{\ldots}{ignored.}
}
\description{
Plots a hierarchical_coverage_regex object
}

