% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage_term.R
\name{plot.hierarchical_coverage_term}
\alias{plot.hierarchical_coverage_term}
\title{Plots a hierarchical_coverage_term Object}
\usage{
\method{plot}{hierarchical_coverage_term}(x, use.terms = nrow(x) <= 30,
  mark.one = FALSE, sort = FALSE, ...)
}
\arguments{
\item{x}{A hierarchical_coverage_term object.}

\item{use.terms}{logical.  If \code{TRUE} terms are plotted on the x axis.
If \code{FALSE} word numbers are.  Te default is to plot terms if they are
equal to or less than 30 in length.}

\item{mark.one}{logical.  If \code{TRUE} a purple horizontal line is added at
100\% and the y axis is extended as well.}

\item{sort}{logical.  If \code{TRUE} the terms are sorted by highest unique
gain.}

\item{\ldots}{ignored.}
}
\description{
Plots a hierarchical_coverage_term object
}

