% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_terms.R
\name{plot.important_terms}
\alias{plot.important_terms}
\title{Plots a important_terms Object}
\usage{
\method{plot}{important_terms}(x, n, as.cloud = FALSE, random.order = FALSE,
  rot.per = 0, ...)
}
\arguments{
\item{x}{The \code{important_terms} object.}

\item{n}{The number of rows to plot.  If integer selects the frequency at
the nth row and plots all rows >= that value.  If proportional (less than 0)
the frequency value for the nth\% row is selected and plots all rows >= that
value.}

\item{as.cloud}{logical.  If \code{TRUE} a wordcloud will be plotted rather
than a bar plot.}

\item{random.order}{logical.  Should the words be place randomly around the
cloud or if \code{FALSE} the more frequent words are in the center of the cloud.}

\item{rot.per}{The precentage of rotated words.}

\item{\ldots}{Other arguments passed to \code{\link[wordcloud]{wordcloud}}.}
}
\description{
Plots a important_terms object.
}
