% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_co_occurrence.R
\name{plot.tag_co_occurrence}
\alias{plot.tag_co_occurrence}
\title{Plots a tag_co_occurrence Object}
\usage{
\method{plot}{tag_co_occurrence}(x, cor = FALSE, edge.weight = 8,
  node.weight = 8, edge.color = "gray80", node.color = "orange",
  bar.color = node.color, font.color = "gray55",
  bar.font.color = ifelse(bar, "gray96", bar.color),
  background.color = NULL, bar.font.size = TRUE, node.font.size = 1.08,
  digits = 1, min.edge.cutoff = 0.15, plot.widths = c(0.65, 0.35),
  bar = TRUE, type = "both", ...)
}
\arguments{
\item{x}{A tag_co_occurrence object.}

\item{cor}{logical.  If \code{TRUE} the correlation matrix is used for the
network graph, otherwise the adjacency matrix is used.}

\item{edge.weight}{A weight for the edges.}

\item{node.weight}{A weight for the nodes.}

\item{edge.color}{A color for the edges.}

\item{node.color}{A color for the nodes.}

\item{bar.color}{A color for the bar fill; defaults to \code{node.color}.}

\item{font.color}{A color for the node and axis text.}

\item{bar.font.color}{A color for the bar/dotplot (mean co-occurrences).}

\item{background.color}{The plot background color.}

\item{bar.font.size}{A font size for the bar/dotplot (mean co-occurrences).
Default tries to calculate based on number of bars.}

\item{node.font.size}{The \code{vertex.label.cex} for the node labels.}

\item{digits}{The number of digits to print for bar/dotplot font (mean
co-occurrences).}

\item{min.edge.cutoff}{A minimum value to use as a cut-off in the network plot.
If a value in the correlation/adjacency matrix is below this value, not egde
will be plotted for the tag (node) connection.}

\item{plot.widths}{A vector of proportions of length 2 and totalling 1
corresponding to the relative width of the network and  bar/dotplot.}

\item{bar}{logical.  If \code{TRUE} a bar plot is used as the second plot,
otherwise a dotplot is used.}

\item{type}{The graph type (network & bar/dotplot).  Choices are:
\code{"bar"}, \code{"network"}, or \code{"both"} corresponding to the graph
type to print.}

\item{\ldots}{Other arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Plots a tag_co_occurrence object
}
\note{
The \code{\link[graphics]{par}} function is called in this function.
This will globally reset the graphics device resulting in odd plots in
the next call to \code{plot}.  If working interactively, restart the graphics
device before plotting again.
}

