% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_before.R
\name{plot.term_loc}
\alias{plot.term_loc}
\title{Plots a term_loc Object}
\usage{
\method{plot}{term_loc}(x, as.cloud = FALSE, random.order = FALSE,
  rot.per = 0, ...)
}
\arguments{
\item{x}{The \code{term_loc} object.}

\item{as.cloud}{logical.  If \code{TRUE} a wordcloud will be plotted rather
than a bar plot.}

\item{random.order}{logical.  Should the words be place randomly around the
cloud or if \code{FALSE} the more frequent words are in the center of the cloud.}

\item{rot.per}{The precentage of rotated words.}

\item{\ldots}{Other arguments passed to \code{\link[wordcloud]{wordcloud}}.}
}
\description{
Plots a term_loc object.
}

