% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_count.R
\name{plot.token_count}
\alias{plot.token_count}
\title{Plots a token_count object}
\usage{
\method{plot}{token_count}(x, labels = FALSE, low = "white",
  high = "red", grid = NA, label.color = "grey70", label.size = 3,
  label.digits = if (weight == "count") {     0 } else {     2 },
  weight = "percent", ...)
}
\arguments{
\item{x}{The token_count object.}

\item{labels}{logical.  If \code{TRUE} the cell count values will be included
on the heatmap.}

\item{low}{The color to be used for lower values.}

\item{high}{The color to be used for higher values.}

\item{grid}{The color of the grid (Use \code{NA} to remove the grid).}

\item{label.color}{The color to make labels if \code{labels = TRUE}.}

\item{label.size}{The size to make labels if \code{labels = TRUE}.}

\item{label.digits}{The number of digits to print if labels are printed.}

\item{weight}{The weight to apply to the cell values for gradient fill.
Currently the following are available:
\code{"proportion"}, \code{"percent"}, and \code{"count"}.  See
\code{\link[termco]{weight}} for additional information.}

\item{\ldots}{ignored}
}
\description{
Plots a token_count object.
}
