% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model.R
\name{plot.validate_model}
\alias{plot.validate_model}
\title{Plots a validate_model Object}
\usage{
\method{plot}{validate_model}(x, digits = 1, size = 0.65, height = 0.3,
  ...)
}
\arguments{
\item{x}{A validate_model object.}

\item{digits}{The number of digits to display n percents.}

\item{size}{The size of error bars.}

\item{height}{The height of error bars.}

\item{\ldots}{ignored.}
}
\description{
Plots a validate_model object
}

