% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ca.R
\name{plot_ca}
\alias{plot_ca}
\title{Plot Term Count as Correspondence Analysis}
\usage{
plot_ca(x, D3 = TRUE, ...)
}
\arguments{
\item{x}{A termco object.}

\item{D3}{logical.  If \code{TRUE} plots in 3-d.}

\item{\ldots}{Other arguments passed to \code{plot} and \code{plot3d.ca}.}
}
\value{
Plots a correspondence analysis.
}
\description{
A wrapper for the \pkg{ca}'s \code{ca} + \code{plot} or \code{plot3d.ca}
functions for plotting a simple correspondence analysis.
}
\examples{
data(presidential_debates_2012)

discoure_markers <- list(
    response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
)

(markers <- with(presidential_debates_2012,
    term_count(dialogue, list(person, time), discoure_markers)
))

plot_ca(markers, D3 = FALSE)
\dontrun{
plot_ca(markers)
}
}
\keyword{ca}
\keyword{correspondence}

