% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Horizontal Bar Plot of Group Counts}
\usage{
plot_counts(x, n = NULL, item.name = "Terms", rev = FALSE, ...)
}
\arguments{
\item{x}{A vector or list of elements.}

\item{n}{Minimum frequency to be shown in the plot.  If \code{NULL} all are
shown.}

\item{item.name}{The name of the variable that contains the groups (different
element in the vector/list).}

\item{rev}{logical.  If \code{TRUE} the bars go from least to greatest.}

\item{\ldots}{ignored.}
}
\value{
\pkg{ggplot2} object.
}
\description{
Plot the counts of groups within a vector or list as a horizontal bar plot.
}
\examples{
x <- sample(LETTERS, 100, TRUE)
y <- lapply(1:100, function(i) sample(LETTERS[1:10], sample(0:5, 1), TRUE))
y <- sapply(y, function(x) {
    if(identical(x, character(0))) return(NULL)
    x
})

plot_count(x)
plot_count(y)
}

