% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cum_percent.R
\name{plot_cum_percent}
\alias{plot_cum_percent}
\title{Plot Cumulative Percent of Terms}
\usage{
plot_cum_percent(x, rotate.term = TRUE, ...)
}
\arguments{
\item{x}{A \code{frequent_terms} object.}

\item{rotate.term}{logical.  If \code{TRUE} the term labels will be rotated 45 degrees.}

\item{\ldots}{ignored.}
}
\description{
Plot a cumulative percentage of terms for frequent terms.
}
\examples{
plot_cum_percent(frequent_terms(presidential_debates_2012[["dialogue"]]))
}
