% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{print.accuracy}
\alias{print.accuracy}
\title{Prints an accuracy Object}
\usage{
\method{print}{accuracy}(x, digits = 3, ...)
}
\arguments{
\item{x}{The accuracy object.}

\item{digits}{The number of digits to print.}

\item{\ldots}{ignored}
}
\description{
Prints an accuracy object
}

