% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_terms.R
\name{print.as_terms}
\alias{print.as_terms}
\title{Prints an as_terms Object}
\usage{
\method{print}{as_terms}(x, ...)
}
\arguments{
\item{x}{The as_terms object.}

\item{\ldots}{ignored}
}
\description{
Prints an as_terms object.
}
