% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_counts.R
\name{print.combine_counts}
\alias{print.combine_counts}
\title{Prints a combine_counts Object}
\usage{
\method{print}{combine_counts}(x, ...)
}
\arguments{
\item{x}{The combine_counts object.}

\item{\ldots}{ignored}
}
\description{
Prints a combine_counts object.
}
