% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{print.coverage}
\alias{print.coverage}
\title{Prints a coverage Object}
\usage{
\method{print}{coverage}(x, ...)
}
\arguments{
\item{x}{The coverage object.}

\item{\ldots}{ignored}

\item{x}{The coverage object.}

\item{\ldots}{ignored}
}
\description{
Prints a coverage object
}

