% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print.evaluate}
\alias{print.evaluate}
\title{Prints an evaluate Object}
\usage{
\method{print}{evaluate}(x, digits = 3, ...)
}
\arguments{
\item{x}{The evaluate object.}

\item{digits}{The number of digits to print.}

\item{\ldots}{ignored}
}
\description{
Prints an evaluate object
}
