% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_terms.R
\name{print.frequent_terms}
\alias{print.frequent_terms}
\title{Prints a frequent_terms Object}
\usage{
\method{print}{frequent_terms}(x, n, ...)
}
\arguments{
\item{x}{The \code{frequent_terms} object.}

\item{n}{The number of rows to print.  If integer selects the frequency at
the nth row and prints all rows >= that value.  If proportional (less than 0)
the frequency value for the nth\% row is selected and prints all rows >= that
value.}

\item{\ldots}{ignored.}
}
\description{
Prints a frequent_terms object.
}
