% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_terms_co_occurrence.R
\name{print.frequent_terms_co_occurrence}
\alias{print.frequent_terms_co_occurrence}
\title{Prints a frequent_terms_co_occurrence Object}
\usage{
\method{print}{frequent_terms_co_occurrence}(x, ...)
}
\arguments{
\item{x}{A frequent_terms_co_occurrence object}

\item{\ldots}{Other arguments passed to \code{\link[termco]{plot.tag_co_occurrence}}.}
}
\description{
Prints a frequent_terms_co_occurrence object
}

