% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{print.hierarchical_coverage}
\alias{print.hierarchical_coverage}
\title{Prints a hierarchical_coverage Object}
\usage{
\method{print}{hierarchical_coverage}(x, ...)
}
\arguments{
\item{x}{The hierarchical_coverage object.}

\item{\ldots}{ignored}
}
\description{
Prints a hierarchical_coverage object
}
