% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage_regex.R
\name{print.hierarchical_coverage_regex}
\alias{print.hierarchical_coverage_regex}
\title{Prints a hierarchical_coverage_regex Object}
\usage{
\method{print}{hierarchical_coverage_regex}(x, ...)
}
\arguments{
\item{x}{A hierarchical_coverage_regex object..}

\item{\ldots}{ignored.}
}
\description{
Prints a hierarchical_coverage_regex object
}
