% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_terms.R
\name{print.important_terms}
\alias{print.important_terms}
\title{Prints an important_terms Object}
\usage{
\method{print}{important_terms}(x, n = NULL, ...)
}
\arguments{
\item{x}{An important_terms object.}

\item{n}{The number of rows to print.  If integer selects the frequency at
the nth row and prints all rows >= that value.  If proportional (less than 0)
the frequency value for the nth\% row is selected and prints all rows >= that
value.}

\item{\ldots}{Ignored.}
}
\description{
Prints an important_terms object
}
