% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_list.R
\name{print.probe_list}
\alias{print.probe_list}
\title{Prints a probe_list Object}
\usage{
\method{print}{probe_list}(x, ...)
}
\arguments{
\item{x}{A probe_list object}

\item{\ldots}{ignored.}
}
\description{
Prints a probe_list object
}

