% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_term.R
\name{print.search_term}
\alias{print.search_term}
\title{Prints a search_term Object}
\usage{
\method{print}{search_term}(x, n = Inf, width = 80,
  sep = "\\n\\n\\n===================================\\n", digits = 5,
  ...)
}
\arguments{
\item{x}{The search_term object.}

\item{n}{The number of elements to print or a range of indexes (greater than 1).}

\item{width}{The width of the printing.}

\item{sep}{A string that separates the elements.}

\item{digits}{The number of coverage digits to print.}

\item{\ldots}{ignored}
}
\description{
Prints a search_term object.
}
