% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{print.split_data}
\alias{print.split_data}
\title{Prints a split_data Object}
\usage{
\method{print}{split_data}(x, n = 6, ...)
}
\arguments{
\item{x}{A split_data object}

\item{n}{Number of elements to print}

\item{\ldots}{ignored.}
}
\description{
Prints a split_data object
}
