% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model.R
\name{print.summary.validate_model}
\alias{print.summary.validate_model}
\title{Prints a summary.validate_model Object}
\usage{
\method{print}{summary.validate_model}(x, digits = 1, ...)
}
\arguments{
\item{x}{A summary.validate_model object.}

\item{digits}{The number of digits to display n percents.}

\item{\ldots}{ignored.}
}
\description{
Prints a summary.validate_model object
}
