% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage_regex.R
\name{print.summary_hierarchical_coverage_regex}
\alias{print.summary_hierarchical_coverage_regex}
\title{Prints a summary_hierarchical_coverage_regex Object}
\usage{
\method{print}{summary_hierarchical_coverage_regex}(x, digits = 1, ...)
}
\arguments{
\item{x}{A summary_hierarchical_coverage_regex object.}

\item{digits}{The number of digits to use in rounding percents.}

\item{\ldots}{ignored.}
}
\description{
Prints a summary_hierarchical_coverage_regex object
}
