% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_term_list.R
\name{print.term_list}
\alias{print.term_list}
\title{Prints a term_list Object}
\usage{
\method{print}{term_list}(x, ...)
}
\arguments{
\item{x}{A \code{term_list} object}

\item{\ldots}{ignored}
}
\description{
Prints a term_list object
}
