% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_list_template.R
\name{print.term_list_template}
\alias{print.term_list_template}
\title{Prints a term_list_template Object}
\usage{
\method{print}{term_list_template}(x, ...)
}
\arguments{
\item{x}{A term_list_template object.}

\item{\ldots}{ignored.}
}
\description{
Prints a term_list_template object
}
