% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_term_list.R
\name{print.termco_nested}
\alias{print.termco_nested}
\title{Prints a termco_nested Object}
\usage{
\method{print}{termco_nested}(x, ...)
}
\arguments{
\item{x}{A \code{termco_nested} object}

\item{\ldots}{ignored}
}
\description{
Prints a termco_nested object
}
