% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_term_list.R
\name{print.termco_unnested}
\alias{print.termco_unnested}
\title{Prints a termco_unnested Object}
\usage{
\method{print}{termco_unnested}(x, ...)
}
\arguments{
\item{x}{A \code{termco_unnested} object}

\item{\ldots}{ignored}
}
\description{
Prints a termco_unnested object
}
