% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_colo_plot_list.R
\name{probe_colo_plot_list}
\alias{probe_colo_plot_list}
\alias{probe_colo_plot}
\title{Generate List of Exploration \code{search_term} + \code{frequent_terms} + \code{plot} Function Calls}
\usage{
probe_colo_plot_list(terms, data.name,
  copy2clip = getOption("termco.copy2clip"), ldots = "")

probe_colo_plot(terms, data, file = "Rplots.pdf", width = 5.5, height = 7,
  ...)
}
\arguments{
\item{terms}{A vector of regex terms to explore (often populated from
\code{frequent_terms}.}

\item{data.name}{A character vector of a data set's name that will serve as
the search context.}

\item{copy2clip}{logical.  If code{TRUE} uses \code{\link[clipr]{write_clip}}
to copy the output to the clipboard.  This option is most useful when trying
to build a list regular expression model for easy pasting between testing
a regex and putting it into the model.  This argument can be set globally by
setting \code{options(termco.copy2clip = TRUE)}.}

\item{ldots}{A string (starting with a comma) of additional arguments to
include in the \code{frequent_terms} function of the list of function calls.}

\item{data}{A vector of character strings.}

\item{file}{A  \file{.pdf} file to plot to.}

\item{width}{The width of the graphics region in inches.}

\item{height}{The height of the graphics region in inches.}

\item{\ldots}{Other arguments passed to frequent terms.}
}
\value{
\code{Vprobe_colo_plot_list} - Returns a string with the concatenated
function calls.  The print method separates the concatenated string into new
line function calls.  If \code{copy2clip = TRUE} the calls are easily pasted
for use in exploration of the terms in the text data set.
}
\description{
\code{probe_colo_plot_list} - The task of determining the regexes used to
feed a \code{term_count} object's \code{term.list} requires careful
exploration of term use in context.  This function generates a list of
function calls for \code{search_term} + \code{frequent_terms} + \code{plot}
with a user predefined data set and term list.  This allows the user to use
bar plot explorations to explore a list of terms (such as from
\code{frequent_terms}) and the accompanying terms that frequently collocate
with these terms.

\code{probe_colo_plot} - Make the plots of \code{probe_colo_plot_list}
directly to an external \file{.pdf} file.
}
\note{
To actually make a \file{.pdf} file of the plots use the
\code{probe_colo_plot} function directly.  Also note that \code{probe_colo_plot_list}
takes a character name for \code{data.name} whereas \code{probe_colo_plot}
takes a an actual vector object for \code{data}.
}
\examples{
probe_colo_plot_list(c("thank", "\\\\bthe", "ee"), "sam_i_am")
probe_colo_plot_list(
    c("thank", "\\\\bthe", "ee"),
    "sam_i_am",
    ldots = ", n = 10, min.char = 5"
)

txt <- presidential_debates_2012[["dialogue"]]
terms <- frequent_terms(txt)[["term"]]
probe_colo_plot_list(terms, "txt")

\dontrun{
probe_colo_list(terms, "txt", copy2clip = TRUE)

#make an external file of plots
probe_colo_plot(terms, txt)
}
}
\seealso{
Other probe functions: \code{\link{probe_colo_list}},
  \code{\link{probe_list}}

Other probe functions: \code{\link{probe_colo_list}},
  \code{\link{probe_list}}
}
