% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_colo_list.R
\name{probe_colo_list}
\alias{probe_colo_list}
\title{Generate List of Exploration \code{search_term} + \code{frequent_terms} Function Calls}
\usage{
probe_colo_list(terms, data.name, copy2clip = getOption("termco.copy2clip"),
  ldots = "")
}
\arguments{
\item{terms}{A vector of regex terms to explore (often populated from
\code{frequent_terms}.}

\item{data.name}{A character vector of a data set's name that will serve as
the search context.}

\item{copy2clip}{logical.  If code{TRUE} uses \code{\link[clipr]{write_clip}}
to copy the output to the clipboard.  This option is most useful when trying
to build a list regular expression model for easy pasting between testing
a regex and putting it into the model.  This argument can be set globally by
setting \code{options(termco.copy2clip = TRUE)}.}

\item{ldots}{A string (starting with a comma) of additional arguments to
include in the \code{frequent_terms} function of the list of function calls.}
}
\value{
Returns a string with the concatenated function calls.  The print
method separates the concatenated string into new line function calls.  If
\code{copy2clip = TRUE} the calls are easily pasted for use in exploration
of the terms in the text data set.
}
\description{
The task of determining the regexes used to feed a \code{term_count} object's
\code{term.list} requires careful exploration of term use in context.  This
function generates a list of function calls for \code{search_term} +
\code{frequent_terms} with a user predefined data set and term list.  This
allows the user to explore a list of terms (such as from \code{frequent_terms})
and the accompanying terms that frequently colocate with these terms.
}
\examples{
probe_colo_list(c("thank", "\\\\bthe\\"", "ee"), "sam_i_am")
probe_colo_list(
    c("thank", "\\\\bthe\\"", "ee"),
    "sam_i_am",
    ldots = ", n = 10, min.char = 5"
)

txt <- presidential_debates_2012[["dialogue"]]
terms <- frequent_terms(txt)[["term"]]
probe_colo_list(terms, "txt")

\dontrun{
probe_colo_list(terms, "txt", copy2clip = TRUE)
}
}
\seealso{
Other probe.functions: \code{\link{probe_colo_plot_list}},
  \code{\link{probe_list}}
}

