% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofastr_reexports.R
\docType{import}
\name{reexports}
\alias{as_dtm}
\alias{as_tdm}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gofastr}{\code{\link[gofastr]{as_dtm}}, \code{\link[gofastr]{as_tdm}}}
}}
\keyword{internal}

