% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_names.R
\name{rename_tags}
\alias{rename_tags}
\title{Rename a term_count Object's Term Columns}
\usage{
rename_tags(x, old, new)
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{old}{A vector of the current names.}

\item{new}{A vector of new names corresponding to the order of \code{old} names.}
}
\value{
Returns a renamed \code{term_count} object.
}
\description{
Safely rename a \code{term_count} object's tag columns and attributes.
}
\examples{
data(presidential_debates_2012)

discoure_markers <- list(
    response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
)

(markers <- with(presidential_debates_2012,
    term_count(dialogue, list(person, time), discoure_markers)
))


rename_tags(markers, old = c('back_channels', 'summons'), new = c('bcs', 's'))
rename_tags(markers, old = c('person'), new = c('people'))
rename_tags(markers, old = c('person', 'back_channels', 'summons'), new = c('people', 'bcs', 's'))
attributes(rename_tags(markers, old = c('back_channels', 'summons'), new = c('bcs', 's')))
}
