% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_colo.R
\name{search_colo}
\alias{search_colo}
\title{Search For Colocations}
\usage{
search_colo(text, colo, n = 10, ...)
}
\arguments{
\item{text}{A text vect to search through.}

\item{colo}{A regular expression to search for (uses \code{grep}).}

\item{\ldots}{Other options related to search_terms() function}
}
\value{
\code{search_colo} - Returns a data.frame containing top terms
colocating with the \code{colo} regex.
}
\description{
\code{search_colo} - Find colocations in text items that contain a term(s).
}
\examples{
search_colo(sam_i_am, "\\\\bsam")
}
\keyword{search}

