% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_term.R
\name{search_term}
\alias{search_term}
\alias{search_term_which}
\title{Search For Terms}
\usage{
search_term(text, term, exclude = NULL, and = NULL, ignore.case = TRUE)

search_term_which(text, term, exclude = NULL, and = NULL,
  ignore.case = TRUE)
}
\arguments{
\item{text}{A text vect to search through.}

\item{term}{A regular expression to search for (uses \code{grep}).}

\item{exclude}{A regular expression to exclude cases for (uses \code{grep}).}

\item{and}{A regular expression that must also be contained in addition to
\code{term} (uses \code{grep}).}

\item{ignore.case}{Logical. Should \code{grep} be done idependant of case?
Can also be a vector of length 3}
}
\value{
\code{search_term} - Returns a text vector meeting \code{term}
regex but not \code{exclude} regex.
}
\description{
\code{search_term} - Find text items that contain a term(s).

\code{search_term_which} - Find index of text items that contain a term(s).
}
\examples{
search_term_which(sam_i_am, "\\\\bsam")
search_term(sam_i_am, "\\\\bsam")
}
\keyword{search}

