% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_term_collocations.R
\name{search_term_collocations}
\alias{search_term_collocations}
\title{Search For Collocations}
\usage{
search_term_collocations(text.var, term, n = 10, ignore.case = TRUE, ...)
}
\arguments{
\item{text.var}{A vector of character strings.}

\item{term}{A regular expression(s) to search for (uses \code{grep}).}

\item{n}{The number of rows to print.  If integer selects the frequency at
the nth row and prints all rows >= that value.  If proportional (less than 0)
the frequency value for the nth\% row is selected and prints all rows >= that
value.}

\item{ignore.case}{logical. Should \code{grep} be done independent of case?}

\item{\ldots}{Other arguments passed to \code{\link[termco]{search_term}}
and \code{\link[termco]{frequent_terms}}.}
}
\value{
Returns a \code{\link[base]{data.frame}} of collocating terms and
frequencies.
}
\description{
A wrapper for \code{\link[termco]{search_term}} +
\code{\link[termco]{frequent_terms}}.  Find words that frequently collocate
with a term(s).  Note that the `\code{term} regexes are eliminated from the
output of top occurring terms.
}
\examples{
## Example 1
search_term_collocations(sam_i_am, "\\\\bsam")
search_term_collocations(sam_i_am, c('green', "\\\\bsam"))
search_term_collocations(sam_i_am, c('green', "\\\\bsam"), min.char=2)

## Example 2
top_colo <- search_term_collocations(
    presidential_debates_2012[["dialogue"]],
    "president",
    n =50
)

top_colo
plot(top_colo)
plot(top_colo, as.cloud=TRUE)

## Example 3
top_colo_exclude <- search_term_collocations(
    presidential_debates_2012[["dialogue"]],
    "president",
    exclude = "obama",
    n =50
)

top_colo_exclude
plot(top_colo_exclude)
}
\author{
Steve T. Simpson and Tyler Rinker <tyler.rinker@gmail.com>.
}
\keyword{collocation}
