% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_meta_tags.R
\name{set_meta_tags}
\alias{set_meta_tags}
\alias{set_meta_tags.data.frame}
\alias{set_meta_tags.list}
\title{Set Meta Tags Attribute on a \code{term_count}/\code{token_count} Object}
\usage{
set_meta_tags(x, tags, ...)

\method{set_meta_tags}{data.frame}(x, tags, ...)

\method{set_meta_tags}{list}(x, tags, ...)
}
\arguments{
\item{x}{A \code{term_count}/\code{token_count} object.}

\item{tags}{A data.frame with a \code{'tags'} column corresponding to the tag
columns in the \code{term_count}/\code{token_count} object.  Can also take a
named list but only allows for tags and meta tags where as the number of tag
levels is unrestricted if passing a data.frame.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{term_count}/\code{token_count} object with the
\code{metatags} attribute set.
}
\description{
Set the \code{metatags} attribute on a \code{term_count}/\code{token_count}
object.
}
\examples{
token_list2 <- list(
    list(
        person = c('sam', 'i')
    ),
    list(
        place = c('here', 'house'),
        thing = c('boat', 'fox', 'rain', 'mouse', 'box', 'eggs', 'ham')
    ),
    list(
        no_like = c('not like'),
        thing = c('train', 'goat'),
        other = c('in')
    ),
    list(
        other = 'i'
    )
)


x <- token_count(sam_i_am, grouping.var = TRUE, token.list = token_list2)
attributes(x)[['metatags']]

tag_list <- list(
    noun = c('person', 'place', 'thing'),
    odd_ones = c('other', 'no_like')
)

library(textshape)

tag_df <- tidy_list(list(
    noun = c('person', 'place', 'thing'),
    odd_ones = c('other', 'no_like')
), 'meta', 'tags')[, 2:1]

x2 <- set_meta_tags(x, tag_list)
attributes(x2)[['metatags']]

x3 <- set_meta_tags(x, tag_df)
attributes(x3)[['metatags']]
}
