% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_coverage_regex.R
\name{summary.hierarchical_coverage_regex}
\alias{summary.hierarchical_coverage_regex}
\title{Summary of an hierarchical_coverage_regex Object}
\usage{
\method{summary}{hierarchical_coverage_regex}(object, ...)
}
\arguments{
\item{object}{An hierarchical_coverage_regex object.}

\item{\ldots}{ignored.}
}
\description{
Summary of an hierarchical_coverage_regex object
}

