% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model.R
\name{summary.validate_model}
\alias{summary.validate_model}
\title{Summary of an validate_model Object}
\usage{
\method{summary}{validate_model}(object, adjust.discrete = FALSE,
  ordered = TRUE, ...)
}
\arguments{
\item{object}{An validate_model object.}

\item{adjust.discrete}{logical.  Should an additional ammount be deducted
from the limits to account for dicrete data}

\item{ordered}{logical.  If \code{TRUE} the rows are ordered by tag accuracy.}

\item{\ldots}{ignored.}
}
\description{
Summary of an validate_model object
}
\references{
\url{http://onlinestatbook.com/2/estimation/proportion_ci.html}
}
