% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_co_occurrence.R
\name{tag_co_occurrence}
\alias{tag_co_occurrence}
\title{Explore Tag Co-Occurrence}
\usage{
tag_co_occurrence(x, ...)
}
\arguments{
\item{x}{A \code{\link[termco]{term_count}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of:
\item{ave_tag}{A 2 column data.frame of tags and the average number of other tags that co-occur with it.}
\item{min_max_cor}{A min-max scaled correlation matrix between tags; diagonals set to 0.}
\item{min_max_adjacency}{A min-max scaled adjacency matrix between tags; diagonals set to 0.}
\item{node_size}{The diagonals from the adjacency matrix; the number of times a tag occurred.}
}
\description{
Explore tag co-occurrence.  The resulting list comes with a plot method that
allows the user to use a network graph to view the connections between
tags as well as the average number of other tags that co-occur with each of
the regex tags.  This can provide information regarding the discriminatory
power of each regex that corresponds to a tag.
}
\examples{
## Example 1
regs <- frequent_terms(presidential_debates_2012[["dialogue"]])[[1]]
regs <- setNames(as.list(regs), regs)

model <- with(presidential_debates_2012,
    term_count(dialogue, TRUE, regs)
)

x <- tag_co_occurrence(model)
names(x)
setNames(
    lapply(names(x), function(a) {if(is.matrix(x[[a]])){ round(x[[a]], 2)} else { x[[a]] }}),
    names(x)
)
heatmap(x[["min_max_cor"]])
heatmap(x[["min_max_adjacency"]])
barplot(sort(x[["node_size"]], TRUE), las=2)
barplot(setNames(x[["ave_tag"]][[2]], x[["ave_tag"]][[1]]), las=2)

## Note. Restart the graphics device between plot calls if working
##       interactively. See `?plot.tag_co_occurrence` for details.
plot(x)
plot(x, cor=FALSE)
plot(x, min.edge.cutoff = .1, node.color = "#1CDB4F")
plot(x, min.edge.cutoff = .2, node.color = "gold", digits = 3)
plot(x, bar = FALSE)

## Example 2
regs2 <- frequent_terms(presidential_debates_2012[["dialogue"]], n=50)[[1]]
regs2 <- setNames(as.list(regs2), regs2)

model2 <- with(presidential_debates_2012,
    term_count(dialogue, TRUE, regs2)
)

x2 <- tag_co_occurrence(model2)
plot(x2)
plot(x2, bar = FALSE, min.edge.cutoff = .13)
plot(x2, bar = FALSE, min.edge.cutoff = .18, node.color = "#ead453")
plot(x2, node.weight = 3)
plot(x2, edge.weight = 20, node.weight = 5)

plot(x2, edge.color = "gray80", node.color = "grey50", font.color = "white",
    background.color = "black")

## Small Number of Tags Example
plot(tag_co_occurrence(markers), node.weight = 30)
}
\author{
Steve T. Simpson and Tyler Rinker <tyler.rinker@gmail.com>.
}

