% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R, R/tag_cols.R
\name{term_cols}
\alias{term_cols}
\alias{tag_cols}
\alias{tag_names}
\alias{group_cols}
\alias{group_names}
\title{Get Tag/Group Columns}
\usage{
term_cols(x, ...)

tag_cols(x, ...)

tag_names(x, ...)

group_cols(x, ...)

group_names(x, ...)
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{tibble} frame of just terms or grouping variables.
}
\description{
Convenience functions to grab just the tag or grouping variable columns from
a \code{term_count} object.
}
\examples{
tag_cols(markers)
group_cols(markers)
}
