% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_before.R
\name{term_before}
\alias{term_after}
\alias{term_before}
\alias{term_first}
\title{Extract Terms from Relative Locations}
\usage{
term_before(text.var, term, ignore.case = TRUE, ...)

term_after(text.var, term, ignore.case = TRUE, ...)

term_first(text.var, ignore.case = TRUE, ...)
}
\arguments{
\item{text.var}{The text string variable.}

\item{term}{A regex term to provide the search position.}

\item{ignore.case}{logical.  If \code{FALSE}, the pattern matching is case
sensitive and if \code{TRUE}, case is ignored during matching.}

\item{\ldots}{ignored.}
}
\value{
Returns a data.frame of terms and frequencies
}
\description{
\code{term_before} - View the frequency of terms before a regex/term.

\code{term_after} - View the frequency of terms after a regex/term.

\code{term_first} - View the frequency of terms starting each string.
}
\examples{
term_before(presidential_debates_2012$dialogue, 'president')
term_after(presidential_debates_2012$dialogue, 'president')
term_after(presidential_debates_2012$dialogue, 'oil')
term_first(presidential_debates_2012$dialogue)

x <- term_before(presidential_debates_2012$dialogue, 'president')
plot(x)

\dontrun{
library(dplyr); library(lexicon)

pos_df_pronouns[['pronoun']][1:5] \%>\%
    lapply(function(x){
        term_after(presidential_debates_2012$dialogue, paste0("\\\\b", x, "\\\\b"))
    }) \%>\%
    setNames(pos_df_pronouns[['pronoun']][1:5])

term_first(presidential_debates_2012$dialogue) \%>\%
    filter(!term \%in\% tolower(sw_dolch) & !grepl("'", term))
}
}

