% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_cols.R
\name{term_cols}
\alias{term_cols}
\alias{group_cols}
\title{Get Term/Group Columns}
\usage{
term_cols(x, ...)

group_cols(x, ...)
}
\arguments{
\item{x}{A \code{term_count} object.}

\item{\ldots}{ignored.}
}
\value{
Returns a \code{tibble} frame of just terms or grouping variables.
}
\description{
Convenience functions to grab just the term or grouping variable columns from
a \code{term_count} object.
}
\examples{
term_cols(markers)
group_cols(markers)
}
