% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncovered.R
\name{uncovered}
\alias{uncovered}
\title{Indices of Uncovered/Untagged Group Variable}
\usage{
uncovered(x, ...)
}
\arguments{
\item{x}{A \code{\link[termco]{term_count}} object.}

\item{\ldots}{ignored.}
}
\value{
Returns logical indeices of untagged/uncovered group variables.
}
\description{
This is most useful when \code{grouping.var = TRUE} and an \code{id} variable
was created that corresponds to the text variable.  This allows the user to
quickly grab the untagged text.
}
\examples{
library(dplyr)

untagged <- presidential_debates_2012 \%>\%
     with(., term_count(dialogue, TRUE, list(manners = c("please|excuse|sorry")))) \%>\%
     uncovered(FALSE) \%>\%
     {presidential_debates_2012[., "dialogue"]} \%>\%
     unlist()

frequent_terms(untagged)
search_term(untagged, colo("romney", "governor"))

search_term(untagged, colo("people")) \%>\%
    frequent_terms()
}

