% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_source_term_list.R
\name{unnest_source_term_list}
\alias{unnest_source_term_list}
\title{Format Term List as Individual Regexes}
\usage{
unnest_source_term_list(file, indices = NULL, ...)
}
\arguments{
\item{file}{Path to external term list.}

\item{indices}{Indices of the elements to retain (used to take part of a
term list).}

\item{\ldots}{ignored.}
}
\value{
Returns an unnested & uncollapsed term list.
}
\description{
For validation purposes it is useful to test an uncollapsed and unnested
term list to validate how well individual regexes are performing without
iterative hierarchical processing.  This function allows the user to read
in an external term list file and format as a term list object made up of the
individual regular expressions for each category.  The regular expression
maintain category names and are numbered sequentially.  This list can then be
modeled and passed to the validation tools to understand regular expression
performance that might not be clear when one builds a larger model containing
many tiers.  After validating, the user can use the information to rearrange
the regular expressions within categories at various tiers depending on the
level of confidence associated with a regex.
}
\examples{
\dontrun{
library(dplyr)
library(textshape)

good_cats <- system.file("termco_docs/categories.R", package = "termco")

good_cats \%>\%
    unnest_source_term_list()

good_cats \%>\%
    unnest_source_term_list() \%>\%
    tidy_list('category', 'regex')
}
}
