% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termco-package.R
\docType{data}
\name{validated}
\alias{validated}
\title{A Simulated \code{validate_model} Output}
\format{A data frame with 57 rows and 2 variables}
\usage{
data(validated)
}
\description{
A dataset containing similated \code{validate_model} output.
}
\details{
\itemize{
  \item tag. The tag.
  \item correct. Dummy coded if the tag correctly assessed the text.
}
}
\keyword{datasets}
