% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_term_list.R
\name{as_term_list}
\alias{as_term_list}
\title{Coerce to Named List}
\usage{
as_term_list(x, add.boundary = FALSE, ...)
}
\arguments{
\item{x}{A vector of strings or a \pkg{quanteda} \code{dictionary}.}

\item{add.boundary}{logical.  If \code{TRUE} a word boundary is place at the
beginning and end of the strings.}

\item{\ldots}{ignored.}
}
\value{
Returns a named list.
}
\description{
Convenience function to convert a data forms of terms into a named list.  For
vectors, names are the same as the terms.
}
\examples{
as_term_list(state.name)

\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse)


x <- presidential_debates_2012[["dialogue"]]

bigrams <- ngram_collocations(x, n=10) \%>\%
    transmute(bigram = paste(term1, term2)) \%>\%
    unlist() \%>\%
    as_term_list()


presidential_debates_2012 \%>\%
    with(term_count(dialogue, person, bigrams))

## dictionary from quanteda
require(quanteda)
mfdict <- dictionary(
    file = "http://ow.ly/VMRkL",
    format = "LIWC"
)

as_term_list(mfdict, TRUE)
}
}
