% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_term_list.R
\name{as_term_list}
\alias{as_term_list}
\title{Convert Vector to Named List}
\usage{
as_term_list(x, ...)
}
\arguments{
\item{x}{A vector of strings.}

\item{\ldots}{ignored.}
}
\value{
Returns a named list.
}
\description{
Convenience function to convert a vector of terms into a named list.  Names
are the same as the terms.
}
\examples{
as_term_list(state.name)

\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse)


x <- presidential_debates_2012[["dialogue"]]

bigrams <- ngram_collocations(x, n=10) \%>\%
    transmute(bigram = paste(term1, term2)) \%>\%
    unlist() \%>\%
    as_term_list()


presidential_debates_2012 \%>\%
    with(term_count(dialogue, person, bigrams))
}
}
