% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colo.R
\name{colo}
\alias{colo}
\title{Make Regex to Locate Strings Containing Co-ocuring Substrings}
\usage{
colo(..., not = NULL, copy2clip = getOption("termco.copy2clip"))
}
\arguments{
\item{not}{A substring to exclude from consideration.}

\item{copy2clip}{logical.  If code{TRUE} uses \code{\link[clipr]{write_clip}}
to copy the output to the clipboard.  This option is most useful when trying
to build a list regular expression model for easy pasting between testing
a regex and putting it into the model.  This argument can be set globally by
setting \code{options(termco.copy2clip = TRUE)}.}

\item{\ldots}{Terms that cooccur/collocate}
}
\value{
Returns a regular expression.  If Windows attempts to copy to
clipboard as well.
}
\description{
Make a regex to locate strings that contain >= 2 substrings with optional
negation.
}
\examples{
\dontrun{
colo('overall', 'course')
colo('overall', 'course', "eval")
colo('overall', 'course', not="instructor")

search_term(sam_i_am, colo("^i\\\\b", "like"))
search_term(sam_i_am, colo("^i\\\\b", "like", "not"))
search_term(sam_i_am, colo("^i\\\\b", "like|not"))
search_term(sam_i_am, colo("^i\\\\b", "like", not="not"))
}
}
\keyword{collocate}
\keyword{cooccur}
