% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Coverage for Various Objects}
\usage{
coverage(x, ...)
}
\arguments{
\item{x}{A logical vector, \code{termc_count}, or \code{search_term} object.}

\item{\ldots}{Ignored.}
}
\value{
\code{termc_count} - Returns a proportion of elements covered by the search.

\code{coverage.term_count} - Returns a list:
\item{not}{A logical vector of all rows not covered (row sums equal zero)}
\item{covered}{A logical vector of all rows covered (row sums greater than zero)}
\item{coverage}{The percentage rate of \deqn{\frac{covered}{not + covered}}{covered/(not + covered)}}
\item{n_covered}{The row sums of the unique terms}
\item{total_terms}{The row sums of the terms}
}
\description{
\code{coverage} - Get coverage of a logical vector, \code{term_count}, or \code{search_term} object.

\code{coverage.term_count} - Extract coverage information from a \code{term_count}
object including the percentage of rows that sum to zero as well as the
location of non-covered rows for easy extraction.
}
\examples{
coverage(sample(c(TRUE, FALSE), 1000, TRUE))

data(presidential_debates_2012)

discoure_markers <- list(
    like = c("love", "like"),
    water = c("lake", "ocean", "water"),
    justify = c("because"),
    he = c("\\\\bhe", "him"),
    we = c("\\\\bwe", "\\\\bus", "\\\\bour")
)

library(dplyr)
(markers2 <- with(presidential_debates_2012,
    term_count(dialogue, TRUE, discoure_markers)
))

coverage(markers2)

presidential_debates_2012[coverage(markers2)$not, "dialogue"] \%>\%
   c()
}
\keyword{coverage}

