% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_terms.R
\name{drop_terms}
\alias{drop_terms}
\title{Drop Terms from a Term List}
\usage{
drop_terms(x, drop.terms, ...)
}
\arguments{
\item{x}{A term list.}

\item{drop.terms}{A vector of terms to drop.}

\item{\ldots}{ignored.}
}
\value{
Returns a term list
}
\description{
Usage allows the user to explore/iterate on a term list and drop terms prior
prior to \code{term_count} use without manually editing an external term list
file.
}
\examples{
## Single level term list
discoure_markers <- list(
    response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
)
drop_terms(discoure_markers, 'response_cries')
drop_terms(discoure_markers, c('summons', 'response_cries'))

## Hierarchical term list
trpl_list <- list(
    list(
        response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
        back_channels = c("uh[- ]huh", "uhuh", "yeah"),
        summons = "hey",
        justification = "because"
    ),
    list(summons ='the'),
    list(summons = 'it', justification = 'ed\\\\s')
)

drop_terms(trpl_list, 'response_cries')
drop_terms(trpl_list, c('summons', 'response_cries'))
drop_terms(trpl_list, c('summons', 'response_cries', 'justification'))
}
