% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_ngrams.R
\name{frequent_ngrams}
\alias{frequent_ngrams}
\title{Ngram Collocations}
\usage{
frequent_ngrams(
  text.var,
  n = 20,
  gram.length = 2:3,
  stopwords = stopwords::stopwords("english"),
  min.char = 4,
  max.char = Inf,
  order.by = "frequency",
  stem = FALSE,
  language = "porter",
  ...
)
}
\arguments{
\item{text.var}{A vector of character strings.}

\item{n}{The number of rows to include.}

\item{gram.length}{The length of ngram to generate (2-3).}

\item{stopwords}{A vector of stopwords to exclude.}

\item{min.char}{The minimum number of characters a word must be (including
apostrophes) for inclusion.}

\item{max.char}{The maximum number of characters a word must be (including
apostrophes) for inclusion.}

\item{order.by}{The name of the measure column to order by: \code{"frequency"},
\code{"lambda"}, \code{"z"}.}

\item{stem}{logical.  If \code{TRUE} the \code{\link[SnowballC]{wordStem}}
is used with \code{language = "porter"} as the default.  Note that stopwords
will be stemmed as well.}

\item{language}{The stem language to use (see  \code{\link[SnowballC]{wordStem}}).}

\item{\ldots}{Other arguments passed to \code{\link[quanteda]{collocations}}.}
}
\value{
Retuns a data.frame of terms and frequencies.
}
\description{
Find a important ngram (2-3) collocations.  Wraps \code{\link[quanteda]{collocations}}
to provide stopword, min/max characters, and stemming with a generic \code{plot}
function.
}
\examples{
\dontrun{
x <- presidential_debates_2012[["dialogue"]]

frequent_ngrams(x)
frequent_ngrams(x, n = 50)
frequent_ngrams(x, stopwords = c(stopwords::stopwords("english"), "american", "governor"))
frequent_ngrams(x, gram.length = 3)
frequent_ngrams(x, gram.length = 3, stem = TRUE)
frequent_ngrams(x, order.by = "lambda")

plot(frequent_ngrams(x))
plot(frequent_ngrams(x, n = 40))
plot(frequent_ngrams(x, order.by = "lambda"))
plot(frequent_ngrams(x, gram.length = 3))
}
\dontrun{
## ngram feature extraction
if (!require("pacman")) install.packages("pacman")
pacman::p_load(termco, dplyr, textshape, magrittr)

ngrams <- presidential_debates_2012 \%$\%
    frequent_ngrams(dialogue, n=10) \%>\%
    pull(collocation) \%>\%
    as_term_list() 


ngram_features <- presidential_debates_2012 \%>\%
    with(term_count(dialogue, person, ngrams)) \%>\%
    as_dtm() 

ngram_features

## tidied features
ngram_features \%>\%
    textshape::tidy_dtm()
}
}
\seealso{
\code{\link[quanteda]{collocations}}
}
\keyword{frequency}
\keyword{term}
\keyword{word}
