% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram_collocations.R
\name{plot.ngram_collocations}
\alias{plot.ngram_collocations}
\title{Plots a ngram_collocations Object}
\usage{
\method{plot}{ngram_collocations}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{The \code{ngram_collocations} object.}

\item{plot}{logical.  If \code{TRUE} the output is plotted.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of the three \pkg{ggplot2} objects that make the
combined plot.
}
\description{
Plots a ngram_collocations object.
}

