% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_co_occurrence.R
\name{plot.tag_co_occurrence}
\alias{plot.tag_co_occurrence}
\title{Plots a tag_co_occurrence Object}
\usage{
\method{plot}{tag_co_occurrence}(x, cor = TRUE, edge.weight = 8,
  node.weight = 8, edge.color = "gray80", node.color = "orange",
  bar.color = node.color, font.color = "gray55",
  bar.font.color = ifelse(bar, "gray96", bar.color), bar.font.size = TRUE,
  digits = 1, min.edge.cutoff = 0.15, plot.widths = c(0.65, 0.35),
  bar = TRUE, ...)
}
\arguments{
\item{x}{A tag_co_occurrence object.}

\item{cor}{logical.  If \code{TRUE} the correlation matrix is used for the
network graph, otherwise the adjacency matrix is used.}

\item{edge.weight}{A weight for the edges.}

\item{node.weight}{A weight for the nodes.}

\item{edge.color}{A color for the edges.}

\item{node.color}{A color for the nodes.}

\item{bar.color}{A color for the bar fill; defaults to \code{node.color}.}

\item{font.color}{A color for the node and axis text.}

\item{bar.font.color}{A color for the bar/dotplot (mean co-occurrences).}

\item{bar.font.size}{A font size for the bar/dotplot (mean co-occurrences).
Default tries to calculate based on number of bars.}

\item{digits}{The number of digits to print for bar/dotplot font (mean
co-occurrences).}

\item{min.edge.cutoff}{A minimum value to use as a cut-off in the network plot.
If a value in the correlation/adjacency matrix is below this value, not egde
will be plotted for the tag (node) connection.}

\item{plot.widths}{A vector of proportions of length 2 and totalling 1
corresponding to the relative width of the network and  bar/dotplot.}

\item{bar}{logical.  If \code{TRUE} a bar plot is used as the second plot,
otherwise a dotplot is used.}

\item{\ldots}{Other arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\description{
Plots a tag_co_occurrence object
}

