% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/term_count.R
\name{plot.term_count}
\alias{plot.term_count}
\title{Plots a term_count object}
\usage{
\method{plot}{term_count}(x, labels = FALSE, group = NULL, low = "white",
  high = "red", grid = NA, label.color = "grey70", label.size = 3,
  weight = "proportion", ...)
}
\arguments{
\item{x}{The term_count object.}

\item{labels}{logical.  If \code{TRUE} the cell count values will be included
on the heatmap.}

\item{group}{A character vector of the group variable names to include.
Defaults to all group variables used in the call to
\code{\link[termco]{term_count}}.}

\item{low}{The color to be used for lower values.}

\item{high}{The color to be used for higher values.}

\item{grid}{The color of the grid (Use \code{NA} to remove the grid).}

\item{label.color}{The color to make labels if \code{labels = TRUE}.}

\item{label.size}{The size to make labels if \code{labels = TRUE}.}

\item{weight}{The weight to apply to the cell values for gradient fill.
Currently the following are available:
\code{"proportion"}, \code{"percent"}.  See \code{\link[termco]{weight}} for
additional information.}

\item{\ldots}{ignored}
}
\description{
Plots a term_count object.
}

