% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Horizontal Bar Plot of Group Counts}
\usage{
plot_counts(x, n = NULL, percent = TRUE, item.name = "Terms",
  rev = FALSE, drop = TRUE, ...)
}
\arguments{
\item{x}{A vector or list of elements.}

\item{n}{Minimum frequency to be shown in the plot.  If \code{NULL} all are
shown.}

\item{percent}{logical.  If \code{TRUE} the x axis is scaled as percentages.
Otherwise, the x axis is counts.}

\item{item.name}{The name of the variable that contains the groups (different
element in the vector/list).}

\item{rev}{logical.  If \code{TRUE} the bars go from least to greatest.}

\item{drop}{logical.  If \code{FALSE} and \code{x} is an \code{as_terms} object
created from a \code{term_count} object, then unfound terms will not be dropped.}

\item{\ldots}{ignored.}
}
\value{
\pkg{ggplot2} object.
}
\description{
Plot the counts of groups within a vector or list as a horizontal bar plot.
}
\examples{
x <- sample(LETTERS, 100, TRUE)
y <- lapply(1:100, function(i) sample(LETTERS[1:10], sample(0:5, 1), TRUE))
y <- sapply(y, function(x) {
    if(identical(x, character(0))) return(NULL)
    x
})

plot_counts(x)
plot_counts(y)

## Example
library(dplyr)
data(presidential_debates_2012)

discoure_markers <- list(
    response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
)

presidential_debates_2012 \%>\%
    with(., term_count(dialogue, TRUE, discoure_markers)) \%>\%
    as_terms() \%>\%
    plot_counts() +
        ggplot2::xlab("Tags")

presidential_debates_2012 \%>\%
    with(., term_count(dialogue, TRUE, discoure_markers)) \%>\%
    as_terms() \%>\%
    plot_freq(size=3) +
        ggplot2::xlab("Number of Tags")

presidential_debates_2012 \%>\%
    with(., term_count(dialogue, TRUE, discoure_markers)) \%>\%
    as_terms() \%>\%
    plot_counts(percent=FALSE, item.name = "Tags")
}
