% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{Vertical Bar Plot of Frequencies of Counts}
\usage{
plot_freq(x, direct.label = TRUE, label.diff, digits = 1,
  top.diff.weight = 0.06, ...)
}
\arguments{
\item{x}{A vector or list of elements.}

\item{direct.label}{logical.  If \code{TRUE} count + percent labels are
placed above bars.}

\item{label.diff}{The amount to place the labels above the bars.  If
\code{missing} a reasonable guess is attempted.}

\item{digits}{The number of percent digits to print.}

\item{top.diff.weight}{A weight to apply to the space between the top bar and
the top of the plot area.}

\item{\ldots}{Other arguments passed to \code{\link[ggplot2]{geom_text}}.}
}
\value{
\pkg{ggplot2} object
}
\description{
Plot the counts of groups within a vector or list as a horizontal bar plot.
}
\examples{
x <- sample(LETTERS, 100, TRUE)
y <- lapply(1:100, function(i) sample(LETTERS[1:10], sample(0:5, 1), TRUE))
y <- sapply(y, function(x) {
    if(identical(x, character(0))) return(NULL)
    x
})

plot_freq(x)
plot_freq(y)

## Example
library(dplyr)
data(presidential_debates_2012)

discoure_markers <- list(
    response_cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
)

presidential_debates_2012 \%>\%
    with(., term_count(dialogue, TRUE, discoure_markers)) \%>\%
    as_terms() \%>\%
    plot_freq(size=3) +
        ggplot2::xlab("Number of Tags")

presidential_debates_2012 \%>\%
    with(., term_count(dialogue, TRUE, discoure_markers)) \%>\%
    as_terms() \%>\%
    plot_counts() +
        ggplot2::xlab("Tags")
}

