% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accuracy.R
\name{print.accuracy}
\alias{print.accuracy}
\title{Prints an accuracy Object}
\usage{
\method{print}{accuracy}(x, ...)
}
\arguments{
\item{x}{The accuracy object.}

\item{\ldots}{ignored}
}
\description{
Prints an accuracy object
}

