% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_term.R
\name{print.search_term}
\alias{print.search_term}
\title{Prints a search_term Object}
\usage{
\method{print}{search_term}(x, ...)
}
\arguments{
\item{x}{The search_term object.}

\item{\ldots}{ignored}
}
\description{
Prints a search_term object.
}
