% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_term_list.R
\name{read_term_list}
\alias{read_term_list}
\alias{source_term_list}
\title{Read-In Term List}
\usage{
read_term_list(path, indices = NULL, term.list, ...)

source_term_list(path, indices = NULL, ...)
}
\arguments{
\item{path}{Path to external term list.}

\item{indices}{Indices of the elements to retain (used to take part of a
term list).}

\item{term.list}{A term list object that can be passed rather than an
external file.}

\item{\ldots}{ignored.}
}
\value{
Returns a formatted term list.
}
\description{
Read-in and format a term list from an external file.  \code{read_term_list}
collapses the terms within categories by default.  To negate this behavior
use \code{source_term_list}.
}
\examples{
good_cats <- system.file("termco_docs/categories.R", package = "termco")
bad_cats <- system.file("termco_docs/mal_categories.R", package = "termco")

read_term_list(good_cats)
source_term_list(good_cats)

\dontrun{
## Throws warnings
read_term_list(bad_cats)

## Using term list object
my_term_list <- list(
## Tier 1
    list(
        Response_Cries = c("\\\\boh", "\\\\bah", "\\\\baha", "\\\\bouch", "yuk"),
        Back_Channels = c("uh[- ]huh", "uhuh", "yeah"),
        Summons = "hey",
        Justification = "because"
    ),

 ## Tier 2
    list(
        Summons = 'the',
        Back_Channels = c('\\\\ber+\\\\b', 'hm+'),
        Empty = c('^\\\\s*$')
    ),

 ## Tier 3
    list(
        Summons = c(),
        Back_Channels = c()
    ),

## Tier 4
    list(
        Summons = 'hey you',
        Justification = 'ed\\\\s',
        Exclamation = c('\\\\!$', '^[^w]{0,5}wow+.{0,4}'),
        Empty = c()
    )
)

read_term_list(term.list = my_term_list)
}

## Writing term list for non-R .json others to use:
\dontrun{
my_term_list \%>\%
    jsonlite::toJSON(pretty=TRUE) \%>\%
    stringi::stri_unescape_unicode() \%>\%
    cat(file = 'testing.json')
}
}
