% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/term_count.R
\name{term_count}
\alias{term_count}
\title{Search For and Count Terms}
\usage{
term_count(text.var, grouping.var = NULL, term.list, ignore.case = TRUE)
}
\arguments{
\item{text.var}{The text string variable.}

\item{grouping.var}{The grouping variable(s).  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{term.list}{A list of named character vectors.}

\item{ignore.case}{logical.  If \code{FALSE}, the pattern matching is case
sensitive and if \code{TRUE}, case is ignored during matching.}
}
\value{
Returns a \code{\link[dplyr]{tbl_df}} object of term counts by
grouping variable.
}
\description{
\code{termco} - Search a string by any number of grouping variables for
categories (themes) of grouped root terms/substrings.
}
\examples{
data(pres_debates2012)

discoure_markers <- lapply(list(
    response_cries = c("oh", "ah", "aha", "ouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
), qdapRegex::bind)

(markers <- with(pres_debates2012,
    term_count(dialogue, list(person, time), discoure_markers)
))
}
\keyword{substring}
\keyword{term}

