% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/term_count.R
\name{term_count}
\alias{as.count}
\alias{term_count}
\title{Search For and Count Terms}
\usage{
term_count(text.var, grouping.var = NULL, term.list, ignore.case = TRUE)

as.count(x, ...)
}
\arguments{
\item{text.var}{The text string variable.}

\item{grouping.var}{The grouping variable(s).  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.}

\item{term.list}{A list of named character vectors.}

\item{ignore.case}{logical.  If \code{FALSE}, the pattern matching is case
sensitive and if \code{TRUE}, case is ignored during matching.}

\item{x}{A \code{term_count} object.}

\item{\ldots}{ignored}
}
\value{
Returns a \code{\link[dplyr]{tbl_df}} object of term counts by
grouping variable.
}
\description{
\code{termco} - Search a string by any number of grouping variables for
categories (themes) of grouped root terms/substrings.

Forces a \code{term_count} object to print as count integers rather than a
pretty integer weighted combination
}
\note{
Note that while a \code{\link[termco]{term_count}} object prints as a
combination of integer counts and weighted (default percent of terms) in
parenthesis the underlying object is actually a \code{\link[dplyr]{tbl_df}}
of integer term/substring counts.  The user can alter a
\code{\link[termco]{term_count}} object to print as integer permanently using
the \code{\link[termco]{as.count}} function.
}
\examples{
data(pres_debates2012)

discoure_markers <- lapply(list(
    response_cries = c("oh", "ah", "aha", "ouch", "yuk"),
    back_channels = c("uh[- ]huh", "uhuh", "yeah"),
    summons = "hey",
    justification = "because"
), qdapRegex::bind)

(markers <- with(pres_debates2012,
    term_count(dialogue, list(person, time), discoure_markers)
))

print(markers, pretty = FALSE)
print(markers, zero.replace = "_")
plot(markers)
plot(markers, labels=TRUE)

# permanently remove pretty printing
(markers2 <- as.count(markers))

# manipulating the output in a dplyr chain
library(dplyr)

pres_debates2012 \%>\%
    with(., term_count(dialogue, list(person, time), discoure_markers)) \%>\%
    as.count()  # removes pretty print method (not necessary to manipulate)

pres_debates2012 \%>\%
    with(., term_count(dialogue, list(person, time), discoure_markers)) \%>\%
    mutate(totals = response_cries + back_channels + summons + justification) \%>\%
    arrange(-totals)
}
\keyword{substring}
\keyword{term}

