% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncovered.R
\name{uncovered}
\alias{uncovered}
\alias{get_uncovered}
\title{Uncovered/Untagged Group Variable}
\usage{
uncovered(x, ...)

get_uncovered(x, ...)
}
\arguments{
\item{x}{A \code{\link[termco]{term_count}} object.}

\item{\ldots}{ignored.}
}
\value{
\code{uncovered} - Returns logical indeices of untagged/uncovered group variables.

\code{get_uncovered} - Returns a vector of uncovered text from the
original data set used to train the model.
}
\description{
\code{uncovered} - Get logical vector of uncovered data from the original
text used to build the model.

\code{get_uncovered} - Get text vector from the uncovered data of the original
text used to build the model.
}
\note{
This is most useful when \code{grouping.var = TRUE} and an \code{id} variable
was created that corresponds to the text variable.  This allows the user to
quickly grab the untagged text.
}
\examples{
library(dplyr)

untagged <- presidential_debates_2012 \%>\%
     with(., term_count(dialogue, TRUE, list(manners = c("please|excuse|sorry")))) \%>\%
     uncovered() \%>\%
     {presidential_debates_2012[., "dialogue"]} \%>\%
     unlist(use.names = FALSE)

## Shorthand equivalent to code chunk above
untagged <- presidential_debates_2012 \%>\%
     with(., term_count(dialogue, TRUE, list(manners = c("please|excuse|sorry")))) \%>\%
     get_uncovered()

frequent_terms(untagged)
search_term(untagged, colo("romney", "governor"))

search_term(untagged, colo("people")) \%>\%
    frequent_terms()
}
