% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{Weight Term Counts from \code{term_count}}
\usage{
weight(x, weight = "percent", ...)
}
\arguments{
\item{x}{A \code{\link[termco]{term_count}} object.}

\item{weight}{A weight to use.   Currently the following are available:
\code{"proportion"}, \code{"percent"}.}

\item{\ldots}{ignored}
}
\value{
Returns a weighted \code{\link[dplyr]{tbl_df}} object of term counts
by grouping variable.
}
\description{
Weight term counts from \code{\link[termco]{term_count}} object.
}
\examples{
library(dplyr)
data(markers)

weight(markers, "percent") \%>\%
    arrange(desc(n.words))

weight(markers, 'proportion')
}
\keyword{weight}

